feature: divide firebase.server.ts into separate module files
