+++ app/routes/admin/debug.tsx
@@ -1,13 +1,7 @@
 import { ActionFunction } from "@remix-run/node";
 import { useSubmit } from "@remix-run/react";
 import { useRef } from "react";
-import {
-  debug_addExtraDataToRevenueDB,
-  debug_addUploadedDate,
-  debug_deleteAllTestRevenueData,
-  debug_fixRevenueDataProviderName,
-  debug_initializeIsDiscountedForSettlements,
-} from "~/services/firebase-debug.server";
+import { debug_addExtraDataToRevenueDB } from "~/services/firebase/firebase-debug.server";
 
 export const action: ActionFunction = async ({ request }) => {
   console.log("request submitted");
