+++ app/routes/admin/revenue-calculate.tsx
@@ -9,7 +9,7 @@ import {
   PartnerRevenueStat,
   PartnerRevenueStatTableMemo,
 } from "~/components/revenue_stat";
-import { getRevenueStats } from "~/services/firebase.server";
+import { getRevenueStats } from "~/services/firebase/firebase.server";
 import { dateToDayStr, dayStrToDate, getTimezoneDate } from "~/utils/date";
 import writeXlsxFile from "write-excel-file";
 import dayPickerStyles from "react-day-picker/dist/style.css";
