+++ app/routes/admin/settlement-manage.tsx
@@ -22,10 +22,7 @@ import {
   SettlementSumItem,
   SettlementSumTable,
 } from "~/components/settlement_sum";
-import {
-  getAllSettlementSum,
-  sendSettlementNoticeEmail,
-} from "~/services/firebase.server";
+import { sendSettlementNoticeEmail } from "~/services/firebase/firebase.server";
 import writeXlsxFile from "write-excel-file";
 import { LoadingOverlay, Space } from "@mantine/core";
 import { BasicModal, ModalButton } from "~/components/modal";
@@ -37,6 +34,7 @@ import {
   numeralMonthToKorean,
 } from "~/utils/date";
 import { requireUser } from "~/services/session.server";
+import { getAllSettlementSum } from "~/services/firebase/settlement.server";
 
 interface EmptySettlementBoxProps extends React.HTMLProps<HTMLDivElement> {}
