+++ app/routes/admin/shipped-list.tsx
@@ -3,22 +3,20 @@ import { PageLayout } from "~/components/page_layout";
 import dayPickerStyles from "react-day-picker/dist/style.css";
 import { useEffect, useMemo, useState } from "react";
 import { Link, useLoaderData, useNavigation } from "@remix-run/react";
-import {
-  DaySelectPopover
-} from "~/components/date";
+import { DaySelectPopover } from "~/components/date";
 import { CommonButton } from "~/components/button";
 import { json, LoaderFunction } from "@remix-run/node";
 import { OrderItem, OrderTable } from "~/components/order";
-import {
-  getAllWaybills,
-  getMonthWaybills,
-  getPartnerProfile,
-  getPartnerWaybills,
-} from "~/services/firebase.server";
+import { getPartnerProfile } from "~/services/firebase/firebase.server";
 import { PossibleSellers, SellerSelect } from "~/components/seller";
 import writeXlsxFile from "write-excel-file";
 import { LoadingOverlay, Space } from "@mantine/core";
 import { dateToDayStr, dayStrToDate, getTimezoneDate } from "~/utils/date";
+import {
+  getAllWaybills,
+  getMonthWaybills,
+  getPartnerWaybills,
+} from "~/services/firebase/waybill.server";
 
 interface EmptySettlementBoxProps extends React.HTMLProps<HTMLDivElement> {}
