+++ app/routes/login.tsx
@@ -10,14 +10,11 @@ import { HeaderBox, MobileHeaderBox } from "~/components/header";
 import { LoadingOverlay } from "@mantine/core";
 import { useMemo, useRef, useState } from "react";
 import { getSignInToken } from "~/services/auth.client";
-import {
-  getFirebaseConfig,
-  isAdmin,
-  isStaff,
-} from "~/services/firebase.server";
+import { isAdmin, isStaff } from "~/services/firebase/firebase.server";
 import { createUserSession, User } from "~/services/session.server";
 import { isMobile } from "~/utils/mobile";
 import { useViewportSize } from "@mantine/hooks";
+import { getFirebaseConfig } from "~/services/firebase/firebaseInit.server";
 
 interface LoginPageProps extends React.HTMLProps<HTMLDivElement> {
   isMobile: boolean;
