+++ app/services/firebase/firebase.server.ts
@@ -1,5 +1,4 @@
 // Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
 import {
   and,
   arrayUnion,
@@ -10,7 +9,6 @@ import {
   getCountFromServer,
   getDoc,
   getDocs,
-  getFirestore,
   limit,
   orderBy,
   query,
@@ -23,13 +21,12 @@ import {
 import {
   deleteObject,
   getDownloadURL,
-  getStorage,
   listAll,
   ref,
   uploadBytes,
 } from "firebase/storage";
 import { PartnerProfile } from "~/components/partner_profile";
-import { dateToDayStr, dayStrToDate, getTimezoneDate } from "~/utils/date";
+import { dateToDayStr, getTimezoneDate } from "~/utils/date";
 import { OrderItem } from "~/components/order";
 import {
   createAuthAccount,
@@ -37,11 +34,10 @@ import {
   updateAuthAccount,
 } from "./firebaseAdmin.server";
 import { emailToId } from "~/utils/account";
-import { sendAligoMessage } from "./aligo.server";
+import { sendAligoMessage } from "../aligo.server";
 import { NoticeItem } from "~/components/notice";
 import { Product } from "~/components/product";
-import { SettlementSumItem } from "~/components/settlement_sum";
-import { sendResendEmail } from "./resend.server";
+import { sendResendEmail } from "../resend.server";
 import {
   PossibleCS,
   PossibleOrderStatus,
@@ -52,37 +48,7 @@ import { LofaSellers, NormalPriceStandardSellers } from "~/components/seller";
 import { DiscountData } from "~/components/discount";
 import { getIdFromTime } from "~/components/date";
 import { ExchangeRefundData } from "~/components/exchange_refund";
-
-// Your web app's Firebase configuration
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
-const firebaseConfig = {
-  apiKey: process.env.FIREBASE_API_KEY,
-  authDomain: process.env.FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.FIREBASE_PROJECT_ID,
-  storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.FIREBASE_APP_ID,
-  measurementId: process.env.FIREBASE_MEASUREMENT_ID,
-};
-
-// Initialize Firebase
-let firebaseApp: any;
-let firestore: any;
-let storage: any;
-
-if (!firebaseApp?.apps.length || !firestore.apps.length) {
-  firebaseApp = initializeApp(firebaseConfig);
-  firestore = getFirestore(firebaseApp);
-  storage = getStorage(firebaseApp);
-}
-
-/**
- * Firebase를 위한 설정입니다. Client side에서 로그인하기 위해 사용합니다.
- * @returns firebaseConfig
- */
-export function getFirebaseConfig() {
-  return firebaseConfig;
-}
+import { firestore, storage } from "./firebaseInit.server";
 
 /**
  * 해당 ID의 계정이 어드민인지 확인합니다.
@@ -288,753 +254,6 @@ export async function deletePartnerProfile({ name }: { name: string }) {
 
   addLog("deletePartnerProfile", { docName: name });
   return firestoreResult;
-}
-
-/**
- * 정산내역을 추가합니다.
- * 추가 후 정산합계 기록도 수정합니다.
- * @param settlements: JSON string of settlement items list
- * @returns
- *
- */
-export async function addSettlements({
-  settlements,
-  monthStr,
-}: {
-  settlements: string;
-  monthStr: string;
-}) {
-  try {
-    const time = new Date().getTime();
-    const data = {
-      json: settlements,
-      updateTime: time,
-    };
-    await setDoc(doc(firestore, `settlements-data-add/${monthStr}`), data);
-    addLog("addSettlements", data);
-    return true;
-  } catch (error: any) {
-    sendAligoMessage({
-      text: `[로파파트너] ${error.message ?? error}`,
-      receiver: "01023540973",
-    });
-    return error.message ?? error;
-  }
-}
-
-/**
- * 해당 정산내역들을 삭제합니다
- * @param monthStr: 월,
- * @param settlements: 대상 정산내역, JSON string of SettlementItem array
- */
-
-export async function deleteSettlements({
-  settlements,
-  monthStr,
-}: {
-  settlements: string;
-  monthStr: string;
-}) {
-  try {
-    const time = new Date().getTime();
-    const data = {
-      json: settlements,
-      updateTime: time,
-    };
-    await setDoc(doc(firestore, `settlements-data-delete/${monthStr}`), data);
-    addLog("deleteSettlements", data);
-    return true;
-  } catch (error: any) {
-    sendAligoMessage({
-      text: `[로파파트너] ${error.message ?? error}`,
-      receiver: "01023540973",
-    });
-    return error.message ?? error;
-  }
-}
-
-/**
- * 해당 정산내역들의 배송비를 0원으로 만듭니다.
- * @param monthStr: 월
- * @param settlements: 대상 정산내역, JSON string of SettlementItem array
- */
-
-export async function deleteSettlementsShippingFee({
-  settlements,
-  monthStr,
-}: {
-  settlements: string;
-  monthStr: string;
-}) {
-  try {
-    const time = new Date().getTime();
-    const data = {
-      json: settlements,
-      updateTime: time,
-    };
-    await setDoc(
-      doc(firestore, `settlements-data-delete-shipping-fee/${monthStr}`),
-      data
-    );
-    addLog("addSettlementsShippingFee", data);
-
-    return true;
-  } catch (error: any) {
-    sendAligoMessage({
-      text: `[로파파트너] ${error.message ?? error}`,
-      receiver: "01023540973",
-    });
-    return error.message ?? error;
-  }
-}
-
-/**
- * 정산을 등록한 월의 목록을 불러옵니다.
- * @param
- * @returns
- *  정산 기록이 있는 월들의 리스트 (XX년 XX월)
- */
-export async function getSettlementMonthes() {
-  const settlementsSnap = await getDocs(collection(firestore, "settlements"));
-  return settlementsSnap.docs.map((doc) => doc.id);
-}
-
-/**
- * 정산 데이터를 불러옵니다
- * @param partnerName: 파트너명, monthStr: 월
- * @returns
- *  Array of SettlementItem
- */
-export async function getSettlements({
-  partnerName,
-  monthStr,
-}: {
-  partnerName: string;
-  monthStr: string;
-}) {
-  const settlementsRef = collection(firestore, `settlements/${monthStr}/items`);
-  const settlementsQuery = query(
-    settlementsRef,
-    where("partnerName", "==", partnerName)
-  );
-  const querySnap = await getDocs(settlementsQuery);
-
-  const resultArray = querySnap.docs
-    .map((doc) => {
-      const data = doc.data();
-      if (data.orderDate) {
-        const date = data.orderDate.toDate();
-        data.orderDate = date;
-      }
-      return data;
-    })
-    .sort((a, b) => {
-      // orderDate가 없는 항목은 맨 앞으로 이동
-      if (!a.orderDate) return -1;
-      if (!b.orderDate) return 1;
-
-      // orderDate가 있는 항목은 날짜 순으로 정렬
-      return a.orderDate - b.orderDate;
-    });
-
-  return resultArray;
-}
-
-/**
- * 정산 합계 정보를 불러옵니다
- * @param partnerName: 파트너명, monthStr: 월
- * @returns
- *  해당 월 파트너의 판매처별 정산 금액 합 (settlement_{판매처이름}, shipping_{판매처이름})
- *  없을 경우 null
- */
-export async function getSettlementSum({
-  partnerName,
-  monthStr,
-}: {
-  partnerName: string;
-  monthStr: string;
-}) {
-  const docRef = doc(
-    firestore,
-    `settlements/${monthStr}/partners`,
-    partnerName
-  );
-  const docSnap = await getDoc(docRef);
-  if (docSnap.exists()) {
-    return docSnap.data();
-  } else {
-    return null;
-  }
-}
-
-/**
- * 모든 파트너의 정산 합계 정보를 불러옵니다
- * @param monthStr: 월
- * @returns
- *  해당 월 모든 파트너의 판매처별 정산 금액 합 ([partnerName, data: {settlement_{판매처이름}, shipping_{판매처이름...}])
- */
-export async function getAllSettlementSum({ monthStr }: { monthStr: string }) {
-  const settlementsRef = collection(
-    firestore,
-    `settlements/${monthStr}/partners`
-  );
-  const querySnap = await getDocs(settlementsRef);
-  const result: SettlementSumItem[] = [];
-  const profilesMap = await getAllPartnerProfiles();
-  for (let i = 0; i < querySnap.docs.length; i++) {
-    const doc = querySnap.docs[i];
-    const partnerProfile = profilesMap.get(doc.id);
-    let brn = "";
-    let bankAccount = "";
-    let businessName = "";
-    let businessTaxStandard = "";
-    if (partnerProfile) {
-      brn = partnerProfile.brn ?? "";
-      bankAccount = partnerProfile.bankAccount ?? "";
-      businessName = partnerProfile.businessName ?? "";
-      businessTaxStandard = partnerProfile.businessTaxStandard ?? "";
-    }
-    result.push({
-      partnerName: doc.id,
-      data: doc.data(),
-      brn: brn,
-      bankAccount: bankAccount,
-      businessName: businessName,
-      businessTaxStandard: businessTaxStandard,
-    });
-  }
-  return result;
-}
-
-/**
- * 오늘 주문서가 공유되었는지를 불러옵니다.
- * @param
- * @returns
- *  정산 기록이 있을 경우 오늘 날짜의 string, 없을 경우 null
- */
-export async function isTodayOrderShared() {
-  const today = dateToDayStr(new Date());
-  const docRef = doc(firestore, "orders", today);
-  const docSnap = await getDoc(docRef);
-  if (docSnap.exists()) {
-    return today;
-  } else {
-    return null;
-  }
-}
-
-/**
- * 주문서 내역을 추가합니다.
- * @param orders: OrderItem[], dayStr: string (XXXX-XX-XX)
- * @returns
- *
- */
-export async function addOrders({
-  orders,
-  dayStr,
-}: {
-  orders: OrderItem[];
-  dayStr: string;
-}) {
-  try {
-    let partnerMap: Map<string, PartnerProfile> = await getAllPartnerProfiles();
-    const phoneList: string[] = [];
-    const phoneOrderCount: Map<string, number> = new Map();
-    let orderBatch = writeBatch(firestore);
-    const time = new Date().getTime();
-
-    for (let i = 0; i < orders.length; i++) {
-      const item = orders[i];
-
-      //연락 돌릴 연락처 && 건수 추가
-      const partner = partnerMap.get(item.partnerName);
-      if (partner == undefined) {
-        throw Error("오류: 존재하지 않는 파트너입니다.");
-      }
-
-      const phone = partner.phone;
-      if (phone.length > 0 && !phoneList.includes(phone)) {
-        phoneList.push(phone);
-        const orderCount = phoneOrderCount.get(phone);
-        if (orderCount == undefined) {
-          phoneOrderCount.set(phone, 1);
-        } else {
-          phoneOrderCount.set(phone, orderCount + 1);
-        }
-      }
-
-      const itemDocName = `${time}_${i}`;
-      let itemDocRef = doc(firestore, `orders/${dayStr}/items`, itemDocName);
-      orderBatch.set(itemDocRef, { ...item, orderSharedDate: dayStr });
-
-      //지연주문건에, 주문건 등록 날짜 (Timestamp) 추가해 해당 아이템 추가
-      // (id를 똑같이 쓰므로 함께 삭제하거나 운송장 기입할 때 같은 id 삭제하면 됨)
-      const date = dayStrToDate(dayStr);
-      const timestamp = Timestamp.fromDate(date);
-      const delayedOrderItem = {
-        ...item,
-        orderSharedDate: dayStr,
-        orderTimestamp: timestamp,
-      };
-      let delayedOrderItemDocRef = doc(
-        firestore,
-        `delayed-orders`,
-        itemDocName
-      );
-      orderBatch.set(delayedOrderItemDocRef, delayedOrderItem);
-
-      if ((i % 200 == 1 && i > 1) || i == orders.length - 1) {
-        await orderBatch.commit();
-        if (i < orders.length) {
-          orderBatch = writeBatch(firestore);
-        }
-      }
-    }
-
-    await setDoc(doc(firestore, `orders/${dayStr}`), {
-      isShared: true,
-    });
-
-    for (let i = 0; i < phoneList.length; i++) {
-      const phone = phoneList[i];
-      const orderCount = phoneOrderCount.get(phone);
-      if (orderCount == undefined) {
-        throw Error("오류: 알리고 발신 과정에서 문제가 발생했습니다.");
-      }
-      const response = await sendAligoMessage({
-        text: `[로파파트너] ${dayStr} 주문이 ${orderCount}건 전달되었습니다. 확인 부탁드립니다.`,
-        receiver: phone,
-      });
-      if (response.result_code != 1) {
-        throw Error(`ALIGO 오류: ${response.message}`);
-      }
-    }
-
-    addLog("addOrders", { dayStr: dayStr, orders: orders.toString() });
-
-    return true;
-  } catch (error: any) {
-    await sendAligoMessage({
-      text: error.message ?? error,
-      receiver: "01023540973",
-    });
-    return error.message ?? error;
-  }
-}
-
-/**
- * 해당 날짜의 모든 주문서 정보를 불러옵니다
- * @param dayStr: 날짜 (XXXX-XX-XX)
- * @returns
- *  Array of OrderItem
- */
-export async function getAllOrders(dayStr: string) {
-  const ordersRef = collection(firestore, `orders/${dayStr}/items`);
-  const ordersQuery = query(ordersRef, orderBy("managementNumber"));
-  const querySnap = await getDocs(ordersQuery);
-  return querySnap.docs.map((doc) => doc.data());
-}
-
-/**
- * 해당 날짜의, 지정된 파트너의 주문서 정보를 불러옵니다
- * @param dayStr: 날짜 (XXXX-XX-XX), partnerName: 파트너 이름
- * @returns
- *  Array of OrderItem
- */
-export async function getPartnerOrders({
-  dayStr,
-  partnerName,
-}: {
-  dayStr: string;
-  partnerName: string;
-}) {
-  const ordersRef = collection(firestore, `orders/${dayStr}/items`);
-
-  const ordersQuery = query(
-    ordersRef,
-    where("partnerName", "==", partnerName),
-    orderBy("managementNumber")
-  );
-  const querySnap = await getDocs(ordersQuery);
-
-  return querySnap.docs.map((doc) => doc.data());
-}
-
-/**
- * 해당 주문서 내역들을 삭제합니다
- * 지연주문건과, 완료운송장도 함께 삭제합니다.
- * @param dayStr: 날짜 (XXXX-XX-XX), orders: 삭제할 주문서 목록
- */
-export async function deleteOrders({
-  dayStr,
-  orders,
-}: {
-  orders: OrderItem[];
-  dayStr: string;
-}) {
-  if (orders.length == 0) {
-    return "오류: 입력된 주문이 없습니다.";
-  }
-
-  try {
-    let deleteBatch = writeBatch(firestore);
-
-    for (let i = 0; i < orders.length; i++) {
-      const item = orders[i];
-
-      //order items에 해당 정산아이템 삭제
-      const ordersRef = collection(firestore, `orders/${dayStr}/items`);
-
-      const idQuery = query(
-        ordersRef,
-        where("amount", "==", item.amount),
-        where("customsCode", "==", item.customsCode),
-        where("deliveryRequest", "==", item.deliveryRequest),
-        where("managementNumber", "==", item.managementNumber),
-        where("optionName", "==", item.optionName),
-        where("orderNumber", "==", item.orderNumber),
-        where("orderer", "==", item.orderer),
-        where("ordererPhone", "==", item.ordererPhone),
-        where("partnerName", "==", item.partnerName),
-        where("phone", "==", item.phone),
-        where("productName", "==", item.productName),
-        where("receiver", "==", item.receiver),
-        where("seller", "==", item.seller),
-        where("shippingCompany", "==", item.shippingCompany),
-        where("waybillNumber", "==", item.waybillNumber),
-        where("zipCode", "==", item.zipCode),
-        limit(1)
-      );
-      const querySnap = await getDocs(idQuery);
-      querySnap.forEach(async (document) => {
-        const docName = document.id;
-        deleteBatch.delete(document.ref);
-
-        //지연주문건에서도 삭제
-        let delayedOrderItemDocRef = doc(firestore, `delayed-orders`, docName);
-        deleteBatch.delete(delayedOrderItemDocRef);
-
-        //완료운송장도 있으면 삭제
-        const waybillSharedDate = document.data().waybillSharedDate;
-
-        if (waybillSharedDate.length > 0) {
-          let waybillItemDocRef = doc(
-            firestore,
-            `waybills/${waybillSharedDate}/items`,
-            docName
-          );
-          deleteBatch.delete(waybillItemDocRef);
-        }
-      });
-      if ((i % 130 == 1 && i > 1) || i == orders.length - 1) {
-        await deleteBatch.commit();
-        if (i < orders.length - 1) {
-          deleteBatch = writeBatch(firestore);
-        }
-      }
-    }
-    addLog("deleteOrders", { dayStr: dayStr, orders: orders.toString() });
-    return true;
-  } catch (error: any) {
-    await sendAligoMessage({
-      text: error.message ?? error,
-      receiver: "01023540973",
-    });
-    return error.message ?? error;
-  }
-}
-
-/**
- * 운송장이 기입된 주문서들을 공유합니다.
- * @param dayStr: 날짜 (XXXX-XX-XX), orders: 운송장 기입된 주문서 목록
- * 1. 해당 주문서의 운송장 내용이 업데이트되고,
- * 2. 그 주문서에 해당하는 지연주문건이 삭제되며,
- * 3. 운송장을 공유한 날짜에 완료운송장 항목을 추가합니다. (수정일 경우 삭제 후 재등록)
- */
-export async function addWaybills({
-  dayStr,
-  orders,
-}: {
-  orders: OrderItem[];
-  dayStr: string;
-}) {
-  if (orders.length == 0) {
-    return "오류: 입력된 운송장이 없습니다.";
-  }
-
-  try {
-    let waybillBatch = writeBatch(firestore);
-
-    let day = new Date();
-    day.setDate(day.getDate() + 1);
-    const nextDayStr = dateToDayStr(day);
-
-    for (let i = 0; i < orders.length; i++) {
-      const item = orders[i];
-
-      //items에 해당 주문서 아이템 찾은 후 택배사, 송장번호 정보 기입
-      const ordersRef = collection(firestore, `orders/${dayStr}/items`);
-
-      //기존 주문서 탐색
-      const idQuery = query(
-        ordersRef,
-        where("amount", "==", item.amount),
-        where("customsCode", "==", item.customsCode),
-        where("deliveryRequest", "==", item.deliveryRequest),
-        where("managementNumber", "==", item.managementNumber),
-        where("optionName", "==", item.optionName),
-        where("orderNumber", "==", item.orderNumber),
-        where("orderer", "==", item.orderer),
-        where("ordererPhone", "==", item.ordererPhone),
-        where("partnerName", "==", item.partnerName),
-        where("phone", "==", item.phone),
-        where("productName", "==", item.productName),
-        where("receiver", "==", item.receiver),
-        where("seller", "==", item.seller),
-        where("zipCode", "==", item.zipCode),
-        limit(1)
-      );
-      const querySnap = await getDocs(idQuery);
-
-      if (querySnap.empty) {
-        return "오류: 입력한 운송장에 해당하는 주문건을 찾을 수 없습니다.";
-      }
-
-      querySnap.forEach(async (document) => {
-        const docName = document.id;
-
-        //기존에 운송장 입력한 기록이 있으면 날짜 확인
-        const prevWaybillSharedDate = document.data().waybillSharedDate;
-
-        //주문서에 운송장 내용 기입
-        waybillBatch.update(document.ref, {
-          shippingCompany: item.shippingCompany,
-          waybillNumber: item.waybillNumber,
-          waybillSharedDate: nextDayStr,
-        });
-
-        //지연주문건에서도 삭제
-        let delayedOrderItemDocRef = doc(firestore, `delayed-orders`, docName);
-        waybillBatch.delete(delayedOrderItemDocRef);
-
-        //다른 날짜에 운송장이 이미 있으면 삭제
-        if (
-          prevWaybillSharedDate !== undefined &&
-          prevWaybillSharedDate.length > 0 &&
-          prevWaybillSharedDate !== nextDayStr
-        ) {
-          let prevWaybillDocRef = doc(
-            firestore,
-            `waybills/${prevWaybillSharedDate}/items`,
-            docName
-          );
-          waybillBatch.delete(prevWaybillDocRef);
-        }
-
-        //완료운송장에 추가
-        let waybillItemDocRef = doc(
-          firestore,
-          `waybills/${nextDayStr}/items`,
-          docName
-        );
-        waybillBatch.set(waybillItemDocRef, {
-          ...item,
-          waybillSharedDate: nextDayStr,
-        });
-      });
-
-      if ((i % 100 == 1 && i > 1) || i == orders.length - 1) {
-        await waybillBatch.commit();
-        if (i < orders.length) {
-          waybillBatch = writeBatch(firestore);
-        }
-      }
-    }
-
-    await setDoc(doc(firestore, `waybills/${nextDayStr}`), {
-      isShared: true,
-    });
-    addLog("addWaybills", { dayStr: dayStr, orders: orders.toString() });
-    return true;
-  } catch (error: any) {
-    await sendAligoMessage({
-      text: error.message ?? error,
-      receiver: "01023540973",
-    });
-    return error.message ?? error;
-  }
-}
-
-/**
- * 해당 날짜의 모든 운송장 정보를 불러옵니다
- * @param dayStr: 날짜 (XXXX-XX-XX)
- * @returns
- *  Array of OrderItem
- */
-export async function getAllWaybills(dayStr: string) {
-  const ordersRef = collection(firestore, `waybills/${dayStr}/items`);
-  const ordersQuery = query(ordersRef, orderBy("managementNumber"));
-  const querySnap = await getDocs(ordersQuery);
-  return querySnap.docs.map((doc) => doc.data());
-}
-
-/**
- * 해당 개월 모든 운송장 정보를 불러옵니다
- * @param dayStr: 개월 (XXXX-XX)
- * @returns
- *  Array of OrderItem
- */
-export async function getMonthWaybills(month: string) {
-  const dataList: DocumentData[] = [];
-  for (let i = 1; i <= 31; i++) {
-    const day = i.toString().padStart(2, "0");
-    const dayStr = month + "-" + day;
-    const ordersRef = collection(firestore, `waybills/${dayStr}/items`);
-    const ordersQuery = query(ordersRef, orderBy("managementNumber"));
-    const querySnap = await getDocs(ordersQuery);
-    querySnap.docs.forEach((doc) => dataList.push(doc.data()));
-  }
-  return dataList;
-}
-
-/**
- * 해당 날짜의, 지정된 파트너의 운송장 정보를 불러옵니다
- * @param dayStr: 날짜 (XXXX-XX-XX), partnerName: 파트너 이름
- * @returns
- *  Array of OrderItem
- */
-export async function getPartnerWaybills({
-  dayStr,
-  partnerName,
-}: {
-  dayStr: string;
-  partnerName: string;
-}) {
-  const ordersRef = collection(firestore, `waybills/${dayStr}/items`);
-
-  const ordersQuery = query(
-    ordersRef,
-    where("partnerName", "==", partnerName),
-    orderBy("managementNumber")
-  );
-  const querySnap = await getDocs(ordersQuery);
-
-  return querySnap.docs.map((doc) => doc.data());
-}
-
-/**
- * 운송장을 수정하여 다시 공유합니다.
- * @param dayStr: 운송장 공유 날짜 (XXXX-XX-XX), waybills: 운송장  목록
- * 1. 해당 운송장 내용이 업데이트되고,
- * 2. 운송장의 원 주문서를 수정합니다.
- */
-export async function editWaybills({
-  dayStr,
-  waybills,
-}: {
-  waybills: OrderItem[];
-  dayStr: string;
-}) {
-  if (waybills.length == 0) {
-    return "오류: 입력된 운송장이 없습니다.";
-  }
-
-  try {
-    let waybillBatch = writeBatch(firestore);
-
-    let nextDay = new Date();
-    nextDay.setDate(nextDay.getDate() + 1);
-    const nextDayStr = dateToDayStr(nextDay);
-
-    for (let i = 0; i < waybills.length; i++) {
-      const item = waybills[i];
-
-      const waybillsRef = collection(firestore, `waybills/${dayStr}/items`);
-
-      //기존 운송장 탐색
-      const idQuery = query(
-        waybillsRef,
-        where("amount", "==", item.amount),
-        where("customsCode", "==", item.customsCode),
-        where("deliveryRequest", "==", item.deliveryRequest),
-        where("managementNumber", "==", item.managementNumber),
-        where("optionName", "==", item.optionName),
-        where("orderNumber", "==", item.orderNumber),
-        where("orderer", "==", item.orderer),
-        where("ordererPhone", "==", item.ordererPhone),
-        where("partnerName", "==", item.partnerName),
-        where("phone", "==", item.phone),
-        where("productName", "==", item.productName),
-        where("receiver", "==", item.receiver),
-        where("seller", "==", item.seller),
-        where("zipCode", "==", item.zipCode),
-        limit(1)
-      );
-      const querySnap = await getDocs(idQuery);
-
-      if (querySnap.empty) {
-        console.log("error: not found");
-        return "오류: 수정을 요청한 운송장을 찾을 수 없습니다.";
-      }
-
-      querySnap.forEach(async (document) => {
-        const docName = document.id;
-
-        //운송장의 원 주문서가 공유된 날짜
-        const orderSharedDate = document.data().orderSharedDate;
-
-        //기존 운송장이 공유된 날짜가 다르면 그 운송장 삭제
-        if (dayStr !== nextDayStr) {
-          waybillBatch.delete(document.ref);
-        }
-
-        //완료운송장에 추가(수정)
-        let waybillItemDocRef = doc(
-          firestore,
-          `waybills/${nextDayStr}/items`,
-          docName
-        );
-
-        waybillBatch.set(waybillItemDocRef, {
-          ...item,
-          waybillSharedDate: nextDayStr,
-        });
-
-        //원 주문서 수정
-
-        let orderItemDocRef = doc(
-          firestore,
-          `orders/${orderSharedDate}/items`,
-          docName
-        );
-
-        waybillBatch.update(orderItemDocRef, {
-          shippingCompany: item.shippingCompany,
-          waybillNumber: item.waybillNumber,
-          waybillSharedDate: nextDayStr,
-        });
-      });
-
-      if ((i % 130 == 1 && i > 1) || i == waybills.length - 1) {
-        await waybillBatch.commit();
-        if (i < waybills.length - 1) {
-          waybillBatch = writeBatch(firestore);
-        }
-      }
-    }
-
-    await setDoc(doc(firestore, `waybills/${nextDayStr}`), {
-      isShared: true,
-    });
-    addLog("editWaybills", { dayStr: dayStr, waybills: waybills.toString() });
-    return true;
-  } catch (error: any) {
-    await sendAligoMessage({
-      text: error.message ?? error,
-      receiver: "01023540973",
-    });
-    return error.message ?? error;
-  }
 }
 
 /**
